# swagger.json
{
  "openapi": "3.0.1",
    "info": {
    "title": "Traccar",
    "version": "4.13",
    "description": "Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).",
    "contact": {
      "name": "Traccar Support",
      "url": "https://www.traccar.org/",
      "email": "support@traccar.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://demo.traccar.org/api",
      "description": "Demo Server 1"
    },
    {
      "url": "https://demo2.traccar.org/api",
      "description": "Demo Server 2"
    },
    {
      "url": "https://demo3.traccar.org/api",
      "description": "Demo Server 3"
    },
    {
      "url": "https://server.traccar.org/api",
      "description": "Subscription Server"
    },
    {
      "url": "http://{host}:{port}/api",
      "description": "Other Server",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "enum": [
            "8082",
            "80"
    ],
          "default": "8082"
        }
      }
    }
  ],
    "security": [
        {
            "basicAuth": []
        }
    ],
...
    "paths": {
...
        "/devices": {
            "get": {
                "summary": "Fetch a list of Devices",
                "description": "Without any params, returns a list of the user's devices",
        "tags": [
          "Devices"
        ],
                "parameters": [
                    {
            "name": "all",
            "in": "query",
            "description": "Can only be used by admins or managers to fetch all entities",
            "schema": {
              "type": "boolean"
            }
                    },
                    {
            "name": "userId",
            "in": "query",
            "description": "Standard users can use this only with their own _userId_",
            "schema": {
              "type": "integer"
            }
                    },
                    {
                        "name" : "id",
                        "in" : "query",
                        "description" : "To fetch one or more devices. Multiple params can be passed like `id=31&id=42`",
            "schema": {
              "type": "integer"
            }
                    },
                    {
                        "name" : "uniqueId",
                        "in" : "query",
                        "description" : "To fetch one or more devices. Multiple params can be passed like `uniqueId=333331&uniqieId=44442`",
            "schema": {
              "type": "string"
            }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
            "content": {
              "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                    "$ref": "#/components/schemas/Device"
                            }
                        }
              }
            }
                    },
                    "400": {
            "description": "No permission",
            "content": {}
                    }
                }
            },
            "post": {
                "summary": "Create a Device",
        "tags": [
          "Devices"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Device"
                    }
            }
          },
          "required": true
        },
                "responses": {
                    "200": {
                        "description": "OK",
            "content": {
              "application/json": {
                        "schema": {
                  "$ref": "#/components/schemas/Device"
                        }
                    }
                }
            }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
    "/devices/{id}/accumulators": {
            "put": {
        "summary": "Update total distance and hours of the Device",
        "tags": [
          "Devices"
        ],
                "parameters": [
                    {
            "name": "id",
            "in": "path",
                        "required": true,
                        "schema": {
              "type": "integer"
                        }
                    }
                ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceAccumulators"
              }
            }
          },
          "required": true
        },
                "responses": {
                    "204": {
            "description": "No Content",
            "content": {}
                    }
        },
        "x-codegen-request-body-name": "body"
            }
        },
...
        "/session": {
            "get": {
                "summary": "Fetch Session information",
        "tags": [
          "Session"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
            "schema": {
                        "type": "string"
                    }
          }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
            "content": {
              "application/json": {
                        "schema": {
                  "$ref": "#/components/schemas/User"
                        }
              }
            }
                    },
                    "404": {
            "description": "Not Found",
            "content": {}
                    }
                }
            },
            "post": {
                "summary": "Create a new Session",
        "tags": [
          "Session"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                        "type": "string"
                    },
                  "password": {
                        "type": "string",
                        "format": "password"
                    }
                }
              }
            }
          },
          "required": true
        },
                "responses": {
                    "200": {
                        "description": "OK",
            "content": {
              "application/json": {
                        "schema": {
                  "$ref": "#/components/schemas/User"
                        }
              }
            }
                    },
                    "401": {
            "description": "Unauthorized",
            "content": {}
                    }
                }
            },
            "delete": {
                "summary": "Close the Session",
        "tags": [
          "Session"
                ],
                "responses": {
                    "204": {
            "description": "No Content",
            "content": {}
                    }
                }
            }
        },
...
        "/attributes/computed": {
            "get": {
                "summary": "Fetch a list of Attributes",
                "description": "Without params, it returns a list of Attributes the user has access to",
        "tags": [
          "Attributes"
        ],
                "parameters": [
                    {
            "name": "all",
            "in": "query",
            "description": "Can only be used by admins or managers to fetch all entities",
            "schema": {
              "type": "boolean"
            }
                    },
                    {
            "name": "userId",
            "in": "query",
            "description": "Standard users can use this only with their own _userId_",
            "schema": {
              "type": "integer"
            }
                    },
                    {
            "name": "deviceId",
            "in": "query",
            "description": "Standard users can use this only with _deviceId_s, they have access to",
            "schema": {
              "type": "integer"
            }
                    },
                    {
            "name": "groupId",
            "in": "query",
            "description": "Standard users can use this only with _groupId_s, they have access to",
            "schema": {
              "type": "integer"
            }
                    },
                    {
            "name": "refresh",
            "in": "query",
            "schema": {
              "type": "boolean"
                    }
          }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
            "content": {
              "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                    "$ref": "#/components/schemas/Attribute"
                            }
                        }
                    }
                }
          }
        }
            },
            "post": {
                "summary": "Create an Attribute",
        "tags": [
          "Attributes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Attribute"
                    }
            }
          },
          "required": true
        },
                "responses": {
                    "200": {
                        "description": "OK",
            "content": {
              "application/json": {
                        "schema": {
                  "$ref": "#/components/schemas/Attribute"
                        }
                    }
                }
            }
        },
        "x-codegen-request-body-name": "body"
      }
    },
        "/attributes/computed/{id}": {
            "put": {
                "summary": "Update an Attribute",
        "tags": [
          "Attributes"
        ],
                "parameters": [
                    {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
                    }
                ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Attribute"
              }
            }
          },
          "required": true
        },
                "responses": {
                    "200": {
                        "description": "OK",
            "content": {
              "application/json": {
                        "schema": {
                  "$ref": "#/components/schemas/Attribute"
                        }
                    }
                }
          }
            },
        "x-codegen-request-body-name": "body"
      },
            "delete": {
                "summary": "Delete an Attribute",
        "tags": [
          "Attributes"
        ],
                "parameters": [
                    {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
                    }
          }
                ],
                "responses": {
                    "204": {
            "description": "No Content",
            "content": {}
                    }
                }
            }
        },
...
    },
  "components": {
    "schemas": {
...
        "User": {
        "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "email": {
            "type": "string"
          },
          "phone": {
                    "type": "string"
                },
                "readonly": {
                    "type": "boolean"
                },
                "administrator": {
                    "type": "boolean"
                },
                "map": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "zoom": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "twelveHourFormat": {
                    "type": "boolean"
                },
                "coordinateFormat": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "expirationTime": {
                    "type": "string",
            "description": "in IS0 8601 format. eg. `1963-11-22T18:30:00Z`",
            "format": "date-time"
                },
                "deviceLimit": {
                    "type": "integer"
                },
                "userLimit": {
                    "type": "integer"
                },
                "deviceReadonly": {
                    "type": "boolean"
                },
                "limitCommands": {
                    "type": "boolean"
                },
                "poiLayer": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
          "attributes": {
            "type": "object",
            "properties": {}
            }
        }
        },
...
        "Device": {
        "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "uniqueId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "lastUpdate": {
                    "type": "string",
            "description": "in IS0 8601 format. eg. `1963-11-22T18:30:00Z`",
            "format": "date-time"
                },
                "positionId": {
                    "type": "integer"
                },
                "groupId": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "geofenceIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
          "attributes": {
            "type": "object",
            "properties": {}
            }
        }
        },
...
      "DeviceAccumulators": {
        "type": "object",
            "properties": {
                "deviceId": {
                    "type": "integer"
                },
                "totalDistance": {
                    "type": "number",
                    "description": "in meters"
          },
          "hours": {
            "type": "number"
                }
            }
        },
...
        "Attribute": {
        "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "attribute": {
                    "type": "string"
                },
                "expression": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "description": "String|Number|Boolean"
                }
            }
        },
...
    },
...
    "securitySchemes": {
        "basicAuth": {
        "type": "http",
        "description": "Basic HTTP authorization with _email_ and _password_",
        "scheme": "basic"
        }
    }
  }
}